Move PermissionActivity's functions to StartActivity
