+++ app/src/main/AndroidManifest.xml
@@ -14,11 +14,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/AppTheme">
 
+        <!-- 実行時パーミッション画面 -->
         <activity
             android:name=".start.StartActivity"
-            android:label="@string/app_name"
-            android:launchMode="singleInstance"
-            android:theme="@android:style/Theme.Translucent.NoTitleBar.Fullscreen">
+            android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -31,16 +31,10 @@
                 <category android:name="android.intent.category.BROWSABLE" />
 
                 <data
-                    android:host="quickecho.bellware.jp"
-                    android:path="/android"
+                    android:host="quickecho.page.link"
                     android:scheme="https" />
             </intent-filter>
         </activity>
-        <!-- 実行時パーミッション画面 -->
-        <activity
-            android:name=".start.PermissionActivity"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenSize"
-            android:label="@string/app_name" />
 
         <!-- メイン画面 -->
         <activity
